+++ legacy/src/model/transformations/questionnaire.spec.jsx
@@ -17,13 +17,6 @@ describe('questionnaire', () => {
       Label: ['questionnaire'],
       Name: 'QUESTIONNA',
       lastUpdatedDate: 'Wed Mar 21 2018 10:23:37 GMT+0100 (CET)',
-      DataCollection: [
-        {
-          id: 'campaign01',
-          uri: 'http://ddi:fr.insee:DataCollection.campaign01',
-          Name: 'Campaign 01',
-        },
-      ],
       genericName: 'QUESTIONNAIRE',
 
       agency: 'fr.insee',
@@ -83,15 +76,12 @@ describe('questionnaire', () => {
     const expected = {
       TargetMode: [''],
       agency: 'fr.insee',
-      campaigns: ['campaign01'],
       final: false,
       id: 'jf0s8i94',
       label: 'questionnaire',
       lastUpdatedDate: 'Wed Mar 21 2018 10:23:37 GMT+0100 (CET)',
       name: 'QUESTIONNA',
-      operation: '',
       owner: 'FAKEPERMISSION',
-      serie: '',
       dynamiqueSpecified: 'Redirections',
       formulaSpecified: 'XPATH',
       articulation: {
@@ -153,13 +143,6 @@ describe('questionnaire', () => {
       id: 'jf0s8i94',
       Label: ['questionnaire'],
       lastUpdatedDate: 'Wed Mar 21 2018 10:23:37 GMT+0100 (CET)',
-      DataCollection: [
-        {
-          id: 'campaign01',
-          uri: 'http://ddi:fr.insee:DataCollection.campaign01',
-          Name: 'Campaign 01',
-        },
-      ],
       TargetMode: [''],
       Name: 'QUESTIONNA',
       flowLogic: 'REDIRECTION',
@@ -167,7 +150,6 @@ describe('questionnaire', () => {
     const output = remoteToState1(remote);
     const expected = {
       TargetMode: [''],
-      campaigns: ['campaign01'],
       dynamiqueSpecified: 'Redirections',
       formulaSpecified: 'XPATH',
       final: false,
@@ -188,9 +170,6 @@ describe('questionnaire', () => {
         label: 'questionnaire',
         agency: 'fr.insee',
         lastUpdatedDate: 'Wed Mar 21 2018 10:37:56 GMT+0100 (CET)',
-        serie: 'serie01',
-        operation: 'operation01',
-        campaigns: ['campaign01'],
         TargetMode: [''],
         ComponentGroup: [],
       };
@@ -272,20 +251,6 @@ describe('questionnaire', () => {
             },
           },
         },
-        campaignsStore: {
-          campaign01: {
-            id: 'campaign01',
-            value: 'campaign01',
-            label: 'Campaign 01',
-            operation: 'operation01',
-          },
-          campaign02: {
-            id: 'campaign02',
-            value: 'campaign02',
-            label: 'Campaign 02',
-            operation: 'operation01',
-          },
-        },
       };
       const output = stateToRemote(state, store);
       const expected = {
@@ -363,13 +328,6 @@ describe('questionnaire', () => {
             id: output.ComponentGroup[0].id,
           },
         ],
-        DataCollection: [
-          {
-            Name: 'Campaign 01',
-            id: 'campaign01',
-            uri: 'http://ddi:fr.insee:DataCollection.campaign01',
-          },
-        ],
         Label: ['questionnaire'],
         Name: 'QUESTIONNA',
         TargetMode: [''],
@@ -406,7 +364,6 @@ describe('questionnaire', () => {
         ComponentGroup: [],
         TargetMode: [''],
         agency: 'fr.insee',
-        campaigns: ['campaign02'],
         dynamiqueSpecified: 'Filtres',
         final: false,
         id: 'kgs4uwzd',
@@ -414,9 +371,7 @@ describe('questionnaire', () => {
         lastUpdatedDate:
           'Tue Oct 27 2020 16:36:25 GMT+0100 (heure normale d’Europe centrale)',
         name: 'QUESTIONNA',
-        operation: 'operation01',
         owner: 'FAKEPERMISSION',
-        serie: 'serie01',
         articulation: {
           items: [
             {
@@ -579,20 +534,6 @@ describe('questionnaire', () => {
             },
           },
         },
-        campaignsStore: {
-          campaign01: {
-            id: 'campaign01',
-            value: 'campaign01',
-            label: 'Campaign 01',
-            operation: 'operation01',
-          },
-          campaign02: {
-            id: 'campaign02',
-            value: 'campaign02',
-            label: 'Campaign 02',
-            operation: 'operation01',
-          },
-        },
       };
       const output = stateToRemote(state, store);
       const expected = {
@@ -658,13 +599,6 @@ describe('questionnaire', () => {
             MemberReference: ['kgs4yv5b', 'kgs4v6eo', 'idendquest'],
             Name: 'PAGE_1',
             id: output.ComponentGroup[0].id,
-          },
-        ],
-        DataCollection: [
-          {
-            Name: 'Campaign 02',
-            id: 'campaign02',
-            uri: 'http://ddi:fr.insee:DataCollection.campaign02',
           },
         ],
         FlowControl: [
@@ -764,13 +698,6 @@ describe('questionnaire', () => {
       Label: ['questionnaire'],
       Name: 'QUESTIONNA',
       lastUpdatedDate: 'Wed Mar 21 2018 10:23:37 GMT+0100 (CET)',
-      DataCollection: [
-        {
-          id: 'campaign01',
-          uri: 'http://ddi:fr.insee:DataCollection.campaign01',
-          Name: 'Campaign 01',
-        },
-      ],
       genericName: 'QUESTIONNAIRE',
 
       agency: 'fr.insee',
@@ -782,15 +709,12 @@ describe('questionnaire', () => {
       jf0s8i94: {
         TargetMode: [''],
         agency: 'fr.insee',
-        campaigns: ['campaign01'],
         final: false,
         id: 'jf0s8i94',
         label: 'questionnaire',
         lastUpdatedDate: 'Wed Mar 21 2018 10:23:37 GMT+0100 (CET)',
         name: 'QUESTIONNA',
-        operation: '',
         owner: 'FAKEPERMISSION',
-        serie: '',
         dynamiqueSpecified: 'Redirections',
         formulaSpecified: 'XPATH',
       },
@@ -804,13 +728,6 @@ describe('questionnaire', () => {
       id: 'jf0s8i94',
       Label: ['questionnaire'],
       lastUpdatedDate: 'Wed Mar 21 2018 10:23:37 GMT+0100 (CET)',
-      DataCollection: [
-        {
-          id: 'campaign01',
-          uri: 'http://ddi:fr.insee:DataCollection.campaign01',
-          Name: 'Campaign 01',
-        },
-      ],
       TargetMode: [''],
       Name: 'QUESTIONNA',
       flowLogic: 'REDIRECTION',
@@ -821,7 +738,6 @@ describe('questionnaire', () => {
     const expected = {
       jf0s8i94: {
         TargetMode: [''],
-        campaigns: ['campaign01'],
         dynamiqueSpecified: 'Redirections',
         formulaSpecified: 'VTL',
         final: false,